tools/validate/requirements.txt
jsonschema>=4.21.0
ruamel.yaml>=0.18.6
